Remove config and logging
This makes it more library like. It's so long since I wrote a lib
that I added a configParser and logging framework, these should
instead be managed by whomever is using the library in their own
application.

I also added a bit more fluff such as CI and a good lint clean up