Data structures for conversations
The Generate API endpoint is not suitable for executing tools. Ollama
only supports tool in the chat endpoint and both Ollama and OpenAI need
the tool output fed back to them to make use of it. We'll need to
converse with the model, but to do that, let's define a couple of data
structures that we can use to drive the conversations including tool
output.