Add options to QueryRelevantDocuments, add a new WithScoreThreshold option
It is useful to either retrieve only the relevant documents or none at
all to avoid tainting the conversation with documents that are not
relevant to add.

To allow that, this patch converts the existing Limit parameter to a
functional option and adds a new one WithScoreThreshold that allows to
set the minimal required cosine distance between the document and the
vectorized query.